refactor: split max-lines pilot files and retire suppressions
